backend/foodgram_api/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
Pillow==9.4.0
drf-extra-fields==3.4.1
djoser==2.1.0
djangorestframework-simplejwt==4.7.2
django-filter==23.1
reportlab==3.6.12
